% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_correlation.R
\name{continuous_correlation}
\alias{continuous_correlation}
\title{A rnbapackage function}
\usage{
continuous_correlation(year)
}
\arguments{
\item{year}{The season from which you want the correlation matrix.}
}
\description{
This function allows you to get the correlation matrix of continuous variables from a given season.
}
\examples{
continuous_correlation()
}
\keyword{continuous}
\keyword{correlation}
\keyword{matrix}
