% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{An rnbapackage function}
\usage{
oldest_player(year)
}
\arguments{
\item{year}{the year of which would you like to know the oldest player}
}
\description{
This function allows you to get the oldest player from a given season.
}
\examples{
oldest_player()
}
\keyword{oldest}
\keyword{player}
